requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.10.4
anyio==3.6.2
apache-airflow==2.5.3
apache-airflow-providers-common-sql==1.4.0
apache-airflow-providers-ftp==3.3.1
apache-airflow-providers-http==4.3.0
apache-airflow-providers-imap==3.1.1
apache-airflow-providers-sqlite==3.3.2
apispec==3.3.2
argcomplete==3.0.8
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
Babel==2.12.1
blinker==1.6.2
cachelib==0.9.0
cattrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
ConfigUpdater==3.1.1
connexion==2.14.2
cron-descriptor==1.2.35
croniter==1.3.14
cryptography==40.0.2
Deprecated==1.2.13
dill==0.3.6
dnspython==2.3.0
docutils==0.19
email-validator==1.3.1
exceptiongroup==1.1.1
fastapi==0.95.1
Flask==2.2.4
Flask-AppBuilder==4.1.4
Flask-Babel==2.0.0
Flask-Caching==2.0.2
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
frozenlist==1.3.3
graphviz==0.20.1
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
idna==3.4
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
join==0.1.1
jsonschema==4.17.3
lazy-object-proxy==1.9.0
linkify-it-py==2.0.0
lockfile==0.12.2
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
mdit-py-plugins==0.3.5
mdurl==0.1.2
modules==1.0.0
multidict==6.0.4
numpy
packaging==23.1
pandas
pathspec==0.9.0
pendulum==2.1.2
pluggy==1.0.0
prison==0.2.1
psutil==5.9.5
pycparser==2.21
pydantic==1.10.7
Pygments==2.15.1
PyJWT==2.6.0
pyrsistent==0.19.3
python-daemon==3.0.1
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==8.0.1
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
pytzdata==2020.1
PyYAML==6.0
requests==2.29.0
requests-toolbelt==0.10.1
rfc3339-validator==0.1.4
rich==13.3.5
scikit-learn
scipy
setproctitle==1.3.2
six==1.16.0
sklearn==0.0.post4
sniffio==1.3.0
SQLAlchemy==1.4.47
SQLAlchemy-JSONField==1.0.1.post0
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
starlette==0.26.1
tabulate==0.9.0
tenacity==8.2.2
termcolor==2.3.0
text-unidecode==1.3
threadpoolctl==3.1.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
uc-micro-py==1.0.1
unicodecsv==0.14.1
urllib3==1.26.15
uvicorn==0.21.1
Werkzeug==2.2.3
wrapt==1.15.0
WTForms==3.0.1
yarl==1.9.2
